gpg: fixup imports
