+++ trezor_agent/__main__.py
@@ -2,8 +2,8 @@
 import argparse
 import functools
 import logging
-import re
 import os
+import re
 import subprocess
 import sys
 import time
